Get One Open/Closed Project and Refactoring Profile Creationg
