+++ functions/functions/handlers/projects/closed.js
@@ -1,0 +1,20 @@
+const { db, admin } = require("../../util/admin");
+
+exports.getOneC = (req, res) => {
+    let projectData = {};
+  return db
+    .doc(`/closed/${req.params.projectId}`)
+    .get()
+    .then((doc) => {
+      if (!doc.exists) {
+        return res.status(404).json({ error: `Project not found` });
+      }
+      projectData = doc.data();
+      projectData.projectId = doc.id;
+      return res.status(200).json(projectData);
+    })
+    .catch((err) => {
+      console.error(err);
+      res.status(500).json({ error: err.code });
+    });
+}
